+++ doc/jsdoc-conf.json
@@ -1,5 +1,5 @@
 {
-  "source": { "include": [ "lib/index.js", "lib/compiler.js" ] },
+  "source": { "include": [ "lib/" ] },
   "plugins": [ "plugins/markdown" ],
   "opts": { "destination": "./doc/" }
 }
